+++ src/FLAMCLP.c
@@ -1998,13 +1998,13 @@ static int siClpPrintWritten(
 {
    int                           i;
    TsHdl*                        psHdl=(TsHdl*)pvHdl;
-   long int s=ftell(pfDoc);
+   size_t s=ftell(pfDoc);
    rewind(pfDoc);
    char* pcPge=malloc(s+1);
    if (pcPge==NULL) {
       return CLPERR(psHdl,CLPERR_SYS,"Allocation of memory for temporary file to print page for argument '%s' failed",psHdl->pcCmd);
    }
-   long int r=fread(pcPge,1,s,pfDoc);
+   size_t r=fread(pcPge,1,s,pfDoc);
    if (r!=s) {
       free(pcPge);
       return CLPERR(psHdl,CLPERR_SYS,"Read of temporary file to print page for command '%s' failed",psHdl->pcCmd);
